+++ opensearchtransport/opensearchtransport.go
@@ -382,7 +382,7 @@ func (c *Client) Perform(req *http.Request) (*http.Response, error) {
 		}
 
 		// Delay the retry if a backoff function is configured
-		if c.retryBackoff != nil {
+		if c.retryBackoff != nil && i < c.maxRetries {
 			var cancelled bool
 			timer := time.NewTimer(c.retryBackoff(i + 1))
 			select {